requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.3.0
qdrant-client>=1.6.0
tqdm>=4.65.0
psycopg[binary]>=3.1.0
python-dotenv>=1.0.0
